Made methods unexported
Refactored test in e2e external and internal tests with access to
unexported methods.
